Correct build issued with a C++ compiler (g++, clang++)
Rename variables name and operator to nonconflicting form with C++.

Add needed casts to convert 'ptr' to 'const ptr' or ther other way
around. Do not change the API of functions as this pollutes the
whole codebase with const.

When the impact is low, change 'char *' to 'const char *'.

Use ANSI function definitions in haxUnixUtil.c.