Remove the global state of allocators and adapt the APIs
Changes in public API:
 * Hax_Memoryp - new opaque type in the public header
 * Hax_DbCkalloc, Hax_DbCkfree, Hax_DbCkrealloc,
   Hax_DumpActiveMemory - gained the "Hax_Memoryp *" argument
 * ckalloc, ckfree, ckrealloc - accepts the "Hax_Memoryp *"
   argument
 * Hax_AssembleCmd, Hax_Concat, Hax_CreateCmdBuf,
   Hax_CreateInterp, Hax_DeleteCmdBuf, Hax_Fork, Hax_Merge,
   Hax_CreateHashEntry, Hax_DeleteHashEntry, Hax_DeleteHashTable,
   Hax_HashStats - gained the "Hax_Interp *interp" argument
 * Hax_CmdDeleteProc - gained the "Hax_Interp *interp" argument
 * Hax_FreeResult - uses now Hax_GetMemoryp in the public macro
 * Hax_CreateMemoryManagement, Hax_GetMemoryp - new functions
 * Hax_HashTable::createProc - gained the "Hax_Interp *interp"
   argument

Allocators used to be global with a singleton state. In the new
model Hax users must allocate the Memory context with the
Hax_CreateMemoryManagement() function and use it for initializing
the Hax interpreter with Hax_CreateInterp().

The Hax_CreateMemoryManagement() function accepts dynamically the
tunable options that used to be hardcoded inside haxCkalloc.c:
 * break_on_malloc
 * trace_on_at_malloc
 * alloc_tracing
 * init_malloced_bodies
 * validate_memory

The Memory context is managed independently from the Interpreter
context. This complicates the usage with one additional call on
an initialization, however it separates the memory management
layer making it easier to manage. In theory multiple interpreters
could use the same memory context and after the termination of
Hax interpreter context, it's possible to dump memory still
managed by the subsystem.

The Memoryp (Hax_Memoryp) context attached to an interpreter
can be retrieved with the Hax_GetMemoryp() function. This is
useful in code that is external to libhax, without the ability
to introspect into the private Interp context. With this change
there is no need to pass the Memoryp pointer to Hax modules, as
they can retrieve it from Hax_Interp always.