haxUnixUtil.c: Replace mktemp() with mkstemp()
mktemp() is considered as insecure by the NetBSD libc.

mktemp() is used twice in haxUnixUtil.c. The first usage is
straightforward mapping of mktemp()+open() into mkstemp().

The 2nd one is a workaround to pacify the warning. The file is
opened twice with different permissions. This usage does not map
into the mkstemp() usage. Replace mktemp() with mkstemp()+close(),
followed by the original usage.