Fix index out of bounds use for value.string in haxVar.c
value.string is in reality a flexible length array inside
a union. As it is a valid code to use it in this way, UBSan
reports this as unspecified code semantics:

haxVar.c:483:30: runtime error: index 33 out of bounds for type 'char [4]'
haxVar.c:481:2: runtime error: index 7 out of bounds for type 'char [4]'
haxVar.c:476:6: runtime error: index 7 out of bounds for type 'char [4]'
haxVar.c:480:24: runtime error: index 8 out of bounds for type 'char [4]'

Rework the code and change the string[4] to [1], as it does
not matter. It removes magic usage of length 4 from string.
The length of value.string is controlled by the length of struct.

Replace usage of value.string[] with ((char *)value.string)[].
The end result is the same, but without triggering Undefined
Behavior.